Adds branch for log & initiative integration w/ new file
